ref: Decouple canonical parser from the registry target
